Run Web Stream tests in Node.js as well
