+++ test/spec/browser-index.js
@@ -8,4 +8,5 @@ import './test-common.js'
 import './test-browser-specific.js'
 import './test-parallel-uploads.js'
 import './test-terminate.js'
+import './test-web-stream.js'
 import './test-end-to-end.js'
