+++ test/spec/node-index.js
@@ -2,4 +2,5 @@ import './test-common.js'
 import './test-node-specific.js'
 import './test-parallel-uploads.js'
 import './test-terminate.js'
+import './test-web-stream.js'
 import './test-end-to-end.js'
